Add optional window_ argument to injectStyles
See https://github.com/fkhadra/react-toastify/issues/902 for more context.
This commit adds an optional argument to the `injectStyles` function to be able to specify in which window to inject the styles. This would be very useful for multi-window applications where the styles need to be injected in all of them.